Be more thorough checking cy.pause()
We now check for member expressions  e.g. `cy.member.pause()` as
well as `cy.member().pause()`. This comes as a side effect of how
we now walk an expression tree.

Walking the expression tree was changed to fix a few cases where
we were accessing values that weren't necessarily available on the
expression 'union'.