Add limit for featureTracking event length
Problem: current definition of `event` column allows 50 characters max.
This results in events not logged at all, when we exceed that limit on
the front-end.

This eslint plugin should give immediate feedback to developers about this limitation.
Until it's fixed on the DB side, we should keep that in mind.

Additionally, we should probably print the event in console as well.

Signed-off-by: Kuba Wolanin <hi@kubawolanin.com>